Two Console Applications
