+++ NetMq_RepReq.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetMq.ReqRep.PeopleService"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetMq.ReqRep.Data.Test", "NetMq.ReqRep.Data.Test\NetMq.ReqRep.Data.Test.csproj", "{0E19D251-37B8-4B99-B67E-665704A99015}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetMq.ReqRep.Client", "NetMq_ReqRep_Client\NetMq.ReqRep.Client.csproj", "{8BFE2327-2378-4BF4-9AA5-4C9601A7C06D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{0E19D251-37B8-4B99-B67E-665704A99015}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0E19D251-37B8-4B99-B67E-665704A99015}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E19D251-37B8-4B99-B67E-665704A99015}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BFE2327-2378-4BF4-9AA5-4C9601A7C06D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BFE2327-2378-4BF4-9AA5-4C9601A7C06D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BFE2327-2378-4BF4-9AA5-4C9601A7C06D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BFE2327-2378-4BF4-9AA5-4C9601A7C06D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
