+++ NetMq_RepReq/NetMq.ReqRep.PeopleService.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{18B11A98-EBAF-44D5-A44C-B3ECB6A3D489}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>NetMq.RepReq.PeopleService</RootNamespace>
-    <AssemblyName>NetMq.RepReq.PeopleService</AssemblyName>
+    <RootNamespace>NetMq.ReqRep.PeopleService</RootNamespace>
+    <AssemblyName>NetMq.ReqRep.PeopleService</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
