+++ NetMq_ReqRep_Client/NetMq.ReqRep.Client.csproj
@@ -50,12 +50,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Config.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Client.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
+    <None Include="bin\Debug\NetMq.ReqRep.Client.exe.config" />
+    <None Include="bin\Debug\NetMq_ReqRep_Client.exe.config" />
+    <None Include="bin\Release\NetMq.ReqRep.Client.exe.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="bin\Debug\AsyncIO.dll" />
+    <Content Include="bin\Debug\AsyncIO.pdb" />
+    <Content Include="bin\Debug\NetMQ.dll" />
+    <Content Include="bin\Debug\NetMQ.pdb" />
+    <Content Include="bin\Debug\NetMq.ReqRep.Client.exe" />
+    <Content Include="bin\Debug\NetMq.ReqRep.Client.pdb" />
+    <Content Include="bin\Debug\NetMQ.xml" />
+    <Content Include="bin\Debug\NetMq_ReqRep_Client.exe" />
+    <Content Include="bin\Debug\NetMq_ReqRep_Client.pdb" />
+    <Content Include="bin\Release\AsyncIO.dll" />
+    <Content Include="bin\Release\AsyncIO.pdb" />
+    <Content Include="bin\Release\NetMQ.dll" />
+    <Content Include="bin\Release\NetMQ.pdb" />
+    <Content Include="bin\Release\NetMq.ReqRep.Client.exe" />
+    <Content Include="bin\Release\NetMq.ReqRep.Client.pdb" />
+    <Content Include="bin\Release\NetMQ.xml" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
